examples/prepack/requirements.txt
pyinstaller
pylzss@git+https://github.com/m1stadev/pylzss@v0.3.6
